{{cookiecutter.project_slug}}/requirements.txt
pip==24
ruff==0.5.5
tox==3.14.0
pytest
hydra-core==1.3.2
omegaconf==2.3.0